// src/app/types/character/about/components/character.effects.model.ts
export interface CharacterEffects {
  savingThrows: string[];          // Бонус к спас броскам
  resistance: string[];            // Сопротивление (Уменьшение урона х2, считается после всех модификаторов)
  immunity: string[];              // Иммунитет
}