// src/app/types/character/about/components/character.proficiency.model.ts
export interface CharacterProficiency {
  skills: string[];     // Владение атрибутивными навыками
  languages: string[];  // Владение языками
  tools: string[];      // Владение инструментами
  weapons: string[];    // Владение оружием
  armor: string[];      // Владение доспехами
}